Use only max speed until it gets close
Opted to only use max speed at first, and not later because it might crash into other cars in other areas. However, as I am writing this, I think we could try and make it go back to the original algorithm when it detects something right in front. And then go full speed again when the road is clear.